Move tests in test
