+++ actions/gls-action/tsconfig.json
@@ -3,6 +3,5 @@
   "compilerOptions": {
     "outDir": "dist"
   },
-  "include": ["src"],
-  "exclude": ["node_modules", "dist", "***/__tests__/**", "**/*.test.ts"]
+  "include": ["src"]
 }
