+++ tsconfig.base.json
@@ -9,5 +9,6 @@
     "skipLibCheck": true,
     "types": ["vite/client"]
   },
-  "include": ["src/**/*"]
+  "include": ["src/**/*"],
+  "exclude": ["node_modules", "dist", "***/test/**", "**/*.test.ts"]
 }
