+++ vitest.config.ts
@@ -2,7 +2,7 @@ import {defineConfig} from "vitest/config";
 
 export default defineConfig({
     test: {
-        include: ["actions/**/src/**/*.{test,spec}.{ts,tsx}"],
+        include: ["actions/**/*.{test,spec}.{ts,tsx}"],
         environment: "node",
         exclude: ['**/dist/**'],
     },
